Made the test consumer & producer truely generic
Modified the test producer and consumer to be able to take keys of any
value rather than just string. Updated the person DTO to have a UUID
identifier than set that as the key in the test. This is a more
realistic test case than a simple random string for a key.